chore: reuse tag and location components
